requirements.txt
requests
facebook-sdk
moviepy
Pillow
schedule
numpy
beautifulsoup4
